Create reusable hook for toggling the Help center.
