+++ client/a8c-for-agencies/components/sidebar/header/help-center.tsx
@@ -1,43 +1,15 @@
-import { HelpCenter } from '@automattic/data-stores';
 import { HelpIcon } from '@automattic/help-center';
 import { Button } from '@wordpress/components';
-import {
-	useDispatch as useDataStoreDispatch,
-	useSelect as useDateStoreSelect,
-} from '@wordpress/data';
 import clsx from 'clsx';
-import type { HelpCenterSelect } from '@automattic/data-stores';
+import useHelpCenter from 'calypso/a8c-for-agencies/hooks/use-help-center';
 
-const HELP_CENTER_STORE = HelpCenter.register();
-
-type Props = {
-	onClick?: () => void;
-};
-
-const SidebarHelpCenter = ( { onClick }: Props ) => {
-	const { show, isMinimized } = useDateStoreSelect( ( select ) => {
-		const store = select( HELP_CENTER_STORE ) as HelpCenterSelect;
-		return {
-			show: store.isHelpCenterShown(),
-			isMinimized: store.getIsMinimized(),
-		};
-	}, [] );
-
-	const { setShowHelpCenter, setIsMinimized } = useDataStoreDispatch( HELP_CENTER_STORE );
-
-	const handleToggleHelpCenter = () => {
-		if ( isMinimized ) {
-			setIsMinimized( false );
-		} else {
-			setShowHelpCenter( ! show );
-		}
-		onClick?.();
-	};
+const SidebarHelpCenter = () => {
+	const { toggleHelpCenter, show } = useHelpCenter();
 
 	return (
 		<>
 			<Button
-				onClick={ handleToggleHelpCenter }
+				onClick={ toggleHelpCenter }
 				className={ clsx( 'sidebar__item-help', {
 					'is-active': show,
 				} ) }
