+++ client/a8c-for-agencies/hooks/use-help-center.ts
@@ -1,0 +1,35 @@
+import { HelpCenter } from '@automattic/data-stores';
+import {
+	useDispatch as useDataStoreDispatch,
+	useSelect as useDateStoreSelect,
+} from '@wordpress/data';
+import type { HelpCenterSelect } from '@automattic/data-stores';
+
+const HELP_CENTER_STORE = HelpCenter.register();
+
+export default function useHelpCenter() {
+	const { show, isMinimized } = useDateStoreSelect( ( select ) => {
+		const store = select( HELP_CENTER_STORE ) as HelpCenterSelect;
+		return {
+			show: store.isHelpCenterShown(),
+			isMinimized: store.getIsMinimized(),
+		};
+	}, [] );
+
+	const { setShowHelpCenter, setIsMinimized } = useDataStoreDispatch( HELP_CENTER_STORE );
+
+	const handleToggleHelpCenter = () => {
+		if ( isMinimized ) {
+			setIsMinimized( false );
+		} else {
+			setShowHelpCenter( ! show );
+		}
+		onClick?.();
+	};
+
+	return {
+		toggleHelpCenter: handleToggleHelpCenter,
+		show,
+		isMinimized,
+	};
+}
